+++ examples/demo/src/grids/legacyGrid/commonComponents/searchForm/component.js
@@ -1,8 +1,0 @@
-import template from './searchForm.html'
-
-export default angular
-  .module('app')
-  .component('gridSearch', {
-    template
-  })
-  .name // .name returns the module name