+++ examples/demo/src/svelte/index.js
@@ -12,6 +12,8 @@ import Growl from './Growl/Index.svelte'
 import Items from './Items/Index.svelte'
 import Lists from './Lists/Index.svelte'
 import Selects from './Selects/Index.svelte'
+import Gridz from './Gridz/Index.svelte'
+import Toolbar from './Toolbar/Index.svelte'
 // import avatarsDemoMod from './avatars'
 
 const appMod = angular.module('svelte.demo.components', [ngSvelteMod])
@@ -52,6 +54,12 @@ appMod
   })
   .component('svelteSelectsIndex', {
     controller: ngSvelteShim(Selects)
+  })
+  .component('svelteGridzIndex', {
+    controller: ngSvelteShim(Gridz)
+  })
+  .component('svelteToolbarIndex', {
+    controller: ngSvelteShim(Toolbar)
   })
 
 export default appMod.name