+++ src/ng/gridz/index.js
@@ -1,10 +1,7 @@
 import grid2Mod from './module'
 import './toolbar'
 import './support'
-import './gridz'
+import './gridz-directive'
 import './list'
-// jqgrid jquery setup
-import './jq/jq.gridz'
-import './jq/jq.formatters'
 
 export default grid2Mod