+++ src/ng/gridz/list/ag-grid-list.js
@@ -1,4 +1,4 @@
-import BaseListCtrl from '../../../../src/ng/gridz/list/BaseListCtrl'
+import BaseListCtrl from './BaseListCtrl'
 // import restStoreApi from '../../store/RestStoreApi'
 import _ from 'lodash'