+++ svelte/index.js
@@ -2,7 +2,11 @@ export { default as Avatar } from './Avatar.svelte';
 export { default as Button } from './Button.svelte';
 export { default as Col } from './Col.svelte';
 export { default as Columns } from './Columns.svelte';
-export * from './Card';
+
+export { default as CardHeader } from './Card/CardHeader.svelte';
+export { default as CardTitle } from './Card/CardTitle.svelte';
+export { default as CardSubtitle } from './Card/CardSubtitle.svelte';
+// export * from './Card';
 export * from './Charts';
 // export * from './Form';
 // export * from './List';